pyproject.toml
[tool.poetry]
name = "yesterdays-wisdom"
version = "0.1.0"
description = ""
authors = ["oscar-adserballe"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
protobuf = "4.21.12"
python-dotenv = "^1.0.1"
google-api-python-client = "2.137.0"
google-auth-oauthlib = "1.2.1"
google-cloud-aiplatform = "1.60.0"
requests = "2.32.3"
google-auth = "^2.32.0"
google-generativeai = "^0.7.2"
pillow = "^10.4.0"
click = "^8.1.7"
tika = "^2.6.0"
pytesseract = "^0.3.13"
pdf2image = "^1.17.0"
colorama = "^0.4.6"
pre-commit = "^4.0.1"
pydantic-ai = "^0.0.14"
aiofiles = "^24.1.0"
psutil = "^6.1.1"
pytest-asyncio = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
